cutre

+++ close_residues.py
@@ -1,49 +1,49 @@
-
-import pymol 
 import argparse
+import os
+import pymol
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--protein', '-pt', help='protein file path', required=True)
 args = parser.parse_args()
 
-
-protein_name=args.protein.split('.')[0]
-print(protein_name)
+protein_name = args.protein.split('.')[0]
 
 pymol.finish_launching()
 cmd.load(args.protein)
 
-#First compute the length 
+# First compute the length
 cmd.select('sele, chain A')
 cmd.select('sele, br. sele')
-all='sele'
-allset= set()
-cmd.iterate(selector.process(all), 'allset.add(resv)')
+all_sel = 'sele'
+all_set = set()
+cmd.iterate(all_sel, 'all_set.add(resv)')
 
-length=len(allset)
+length = len(all_set)
 
-filename=f'{protein_name}_residues.txt'
-
-with open(filename,'w') as file:
-    file.write(f'{length}\n')
+# Select interacting residues
 cmd.delete('sele')
-cmd.select(f'sele, chain A w. 4 of chain B')
+cmd.select('sele, chain A w. 4 of chain B')
 cmd.select('sele, br. sele')
-residues='sele'
+residues_sel = 'sele'
 
-resset= set()
-cmd.iterate(selector.process(residues), 'resset.add(resv)')
+res_set = set()
+cmd.iterate(residues_sel, 'res_set.add(resv)')
 
-print(resset)
+# Save data to CSV
+filename = f'close_residues.csv'
+file_exists = os.path.isfile(filename)
 
-with open(filename,'a') as file:
-    for element in resset:
-        file.write(f'{int(element)}\n')
+if file_exists:
+    with open(filename, 'a') as file:
+        # Convert the list of residues to a whitespace-separated string representation
+        res_str = ' '.join(map(str, list(res_set)))
+        file.write(f'{protein_name},{length},{res_str}\n')
+else:
+    with open(filename, 'w') as file:
+        # Write column headers
+        file.write('protein_name,length,interacting_residues\n')
+        # Convert the list of residues to a whitespace-separated string representation
+        res_str = ' '.join(map(str, list(res_set)))
+        file.write(f'{protein_name},{length},{res_str}\n')
 
 pymol.cmd.quit()
-
-
-
-
-
-
